requirements.txt
Django==4.2.11
python-dotenv==1.0.0
Pillow>=10.3.0
django-extensions==3.2.3
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
django-crispy-forms>=2.3
crispy-bootstrap5==2025.6
python-docx==1.1.0
openpyxl==3.1.2
